1016 update

+++ sb1015_2/src/main/java/edu/du/sb1015_2/repository/MyDataRepository.java
@@ -4,6 +4,9 @@ import edu.du.sb1015_2.entity.MyData;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface MyDataRepository extends JpaRepository<MyData, Integer> {
+public interface MyDataRepository extends JpaRepository<MyData, Long> {
+
 }
